fixed merge conflicts in components.js
