+++ packages/transaction/configuration/context.js
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+const TransactionDetailsContext = createContext(
+  {
+    transaction: {},
+    account: {},
+  },
+);
+
+export default TransactionDetailsContext;
