+++ packages/transaction/detail/info/transactionDetailsComponents/transactionVotes.js
@@ -3,11 +3,11 @@ import React, { useContext, useEffect } from 'react';
 import VoteItem from '@transaction/list/row/voteItem';
 import withData from '@common/utilities/withData';
 import { getDelegates } from '@dpos/utilities/api';
-import { Context } from '../../../../views/screens/managers/transactionDetails/transactionDetails';
+import TransactionDetailsContext from '../../../configuration/context';
 import styles from './styles.css';
 
 export const TransactionVotesComp = ({ t, votedDelegates }) => {
-  const { transaction } = useContext(Context);
+  const { transaction } = useContext(TransactionDetailsContext);
   const { votes } = transaction.asset;
 
   useEffect(() => {
