+++ packages/transaction/detail/manager/transactionDetails/index.js
@@ -1,0 +1,37 @@
+/* istanbul ignore file */
+import { compose } from 'redux';
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
+import { getActiveTokenAccount } from '@wallet/utilities/account';
+import { getTransaction } from '@transaction/utilities/api';
+import withData from '@common/utilities/withData';
+import { parseSearchParams } from '@screens/router/searchParams';
+import { withTranslation } from 'react-i18next';
+import TransactionDetailsManager from './transactionDetailsManager';
+
+const mapStateToProps = (state, ownProps) => ({
+  account: getActiveTokenAccount(state),
+  id: ownProps.match.params.id,
+  activeToken: state.settings.token?.active ?? 'LSK',
+});
+
+const apis = {
+  transaction: {
+    apiUtil: (network, { token, transactionId }) =>
+      getTransaction({ network, params: { transactionId } }, token),
+    getApiParams: (state, ownProps) => ({
+      token: state.settings.token.active,
+      transactionId: parseSearchParams(ownProps.location.search).transactionId,
+      network: state.network,
+    }),
+    transformResponse: response => response.data[0] || {},
+    autoload: true,
+  },
+};
+
+export default compose(
+  withRouter,
+  connect(mapStateToProps),
+  withData(apis),
+  withTranslation(),
+)(TransactionDetailsManager);
