+++ packages/views/screens/managers/transactionDetails/index.js
@@ -1,37 +1,46 @@
-/* istanbul ignore file */
-import { compose } from 'redux';
-import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
-import { getActiveTokenAccount } from '@wallet/utilities/account';
-import { getTransaction } from '@transaction/utilities/api';
-import withData from '@common/utilities/withData';
-import { parseSearchParams } from '@screens/router/searchParams';
-import { withTranslation } from 'react-i18next';
-import TransactionDetails from './transactionDetails';
+import React, { useContext, useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { isEmpty } from '@common/utilities/helpers';
+import Box from '@basics/box';
+import BoxHeader from '@basics/box/header';
+import BoxContent from '@basics/box/content';
+import TransactionDetailsContext from '@transaction/configuration/context';
+import NotFound from './notFound';
+import styles from './transactionDetails.css';
+import LayoutSchema from './layoutSchema';
 
-const mapStateToProps = (state, ownProps) => ({
-  account: getActiveTokenAccount(state),
-  id: ownProps.match.params.id,
-  activeToken: state.settings.token?.active ?? 'LSK',
-});
+const TransactionDetails = ({ title }) => {
+  const {
+    schema,
+    error,
+    isLoading,
+    transaction,
+    containerStyle,
+  } = useContext(TransactionDetailsContext);
+  const { t } = useTranslation();
+  const isDataEmpty = useMemo(() => isEmpty(transaction), [transaction]);
 
-const apis = {
-  transaction: {
-    apiUtil: (network, { token, transactionId }) =>
-      getTransaction({ network, params: { transactionId } }, token),
-    getApiParams: (state, ownProps) => ({
-      token: state.settings.token.active,
-      transactionId: parseSearchParams(ownProps.location.search).transactionId,
-      network: state.network,
-    }),
-    transformResponse: response => response.data[0] || {},
-    autoload: true,
-  },
+  if (!error && isDataEmpty) {
+    return <div />;
+  }
+  if (error && isDataEmpty) {
+    return <NotFound t={t} />;
+  }
+
+  const Layout = LayoutSchema[schema ?? transaction.moduleAssetId] || LayoutSchema.default;
+
+  return (
+    <Box isLoading={isLoading} className={`${styles.container} ${containerStyle}`}>
+      {title && (
+        <BoxHeader>
+          <h1>{title}</h1>
+        </BoxHeader>
+      )}
+      <BoxContent className={`${styles.mainContent} ${Layout.className}`}>
+        {Layout.components.map((Component, index) => <Component key={index} t={t} />)}
+      </BoxContent>
+    </Box>
+  );
 };
 
-export default compose(
-  withRouter,
-  connect(mapStateToProps),
-  withData(apis),
-  withTranslation(),
-)(TransactionDetails);
+export default TransactionDetails;
