+++ packages/views/screens/managers/transactionDetailsModal/index.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import Dialog from '@basics/dialog/dialog';
+import TransactionDetailsManager from '../../../../transaction/detail/manager/transactionDetails';
 import TransactionDetails from '../transactionDetails';
 import styles from './styles.css';
 
-const TransactionDetailsModal = (props) => (
+const TransactionDetailsModal = () => (
   <Dialog hasClose className={`${grid.row} ${grid['center-xs']} ${styles.container}`}>
-    <TransactionDetails {...props} title="Transaction details" />
+    <TransactionDetailsManager>
+      <TransactionDetails title="Transaction details" />
+    </TransactionDetailsManager>
   </Dialog>
 );
